Add invisible as default visibility to vault manager button
DropInActivity checks to see if vaultManager was passed true and sets the
vaultManager button to visible, but does not setthe button to invisible
if false or nothing was passed.  However, Androidviews default to visible
which means that the button was always visible.  This sets a default
visibility of invisible which hides the button unless vaultManager is
passed with true.