Fix test expiry date and refactor test client token fixtures
The expiry date helper const is no longer valid, and was preventing
tests from correctly running or testing the appropriate failures via
Robolectric.  Bumping this to a considerable date in the future should
be sufficient.

The refactor for the client token fixtures is due to a bugfix we
deployed [here](https://github.braintreeps.com/team-sdk/braintree-android/pull/446/files) that corrects an unintended behavior
where a merchant could pass a regular string as an auth.  This
reinforces the need to only pass encoded client token or UATs, or
tokenization keys, and rejects random strings as incorrectly formatted.